Reorganize files
There is no big point to split the functionality into such tiny
pieces:

put all matrix specific operations into matrix.c module;
put main processing into processing.c module (it is just 2 functions
at the moment, read_file() and calculate()).

No funtional changes.
