+++ main.c
@@ -21,11 +21,9 @@
  * USA
  */
 
-#include "functions.h"
+#include "processing.h"
 
-#include <ctype.h>
 #include <errno.h>
-#include <math.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
