+++ matrix.h
@@ -1,7 +1,7 @@
-#include <stdlib.h>
-#include <stdio.h>
+#ifndef _MATRIXLAB_MATRIX_H_
+#define _MATRIXLAB_MATRIX_H_
 
-#define MAXNOP 20               /*Max number of operations allowed */
+#include <stdlib.h>
 
 struct m {
     size_t row;
@@ -11,11 +11,10 @@ struct m {
 
 struct m add(const struct m *A, const struct m *B, double n);
 struct m multiply(const struct m *A, const struct m *B);
-void f(double x);
 void print_matrix(struct m *A);
 void transpose(struct m *A);
 double determinant(size_t n, struct m *A);
 void scalar_product(double scalar, struct m *B);
 void inverse(size_t n, struct m *A);
-void read_file(int maxc, FILE * fp);
-void calculate(struct m *matrix, int nop, int id, char op[MAXNOP]);
+
+#endif /* _MATRIXLAB_MATRIX_H_ */
