+++ processing.h
@@ -1,0 +1,8 @@
+#ifndef _MATRIXLAB_PROCESSING_H_
+#define _MATRIXLAB_PROCESSING_H_
+
+#include <stdio.h>
+
+void read_file(int maxc, FILE *fp);
+
+#endif /* _MATRIXLAB_PROCESSING_H_ */
